Modification de l'affichage des salles en fonction du type utilisateur

+++ pages/affichageSalle.php
@@ -78,15 +78,20 @@ $tabLogiciels = listeDesLogiciels();
         <?php endif; ?>
 
         <!-- 1ère ligne avec le bouton "Ajouter" -->
-        <div class="row mb-3">
-            <div class="col-12 text-center text-md-end">
-                <button class="btn-bleu rounded-2" onclick="window.location.href='creationSalle.php';">
-                    <span class="fa-plus">
-                     Ajouter
-                    </span>
-                </button>
-            </div>
-        </div>
+        <?php
+        if($_SESSION['typeUtilisateur'] === 1 ){
+            echo '<div class="row mb-3">
+                        <div class="col-12 text-center text-md-end">
+                            <button class="btn-bleu rounded-2" onclick="window.location.href="creationSalle.php";">
+                                <span class="fa-plus">
+                                    Ajouter
+                                </span>
+                            </button>
+                        </div>
+                  </div>';
+        }
+        ?>
+
 
         <div class="row g-1 justify-content-start"> <!-- Grande row des filtres avec espacement réduit -->
             <!-- Nom des salles -->
@@ -180,7 +185,11 @@ $tabLogiciels = listeDesLogiciels();
                         <th>Type</th>
                         <th>Logiciels</th>
                         <th>Imprimante</th>
-                        <th>Actions</th>
+                        <?php
+                        if($_SESSION['typeUtilisateur'] === 1){
+                            echo '<th>Actions</th>';
+                        }
+                        ?>
                     </tr>
                     </thead>
                     <tbody>
@@ -210,26 +219,27 @@ $tabLogiciels = listeDesLogiciels();
                             echo "<td>".$salle['logiciels']."</td>";
                             // Imprimante : Condition pour afficher Oui/Non
                             echo "<td>".($salle['imprimante'] == 1 ? "Oui" : "Non")."</td>";
+                            if($_SESSION['typeUtilisateur'] === 1){
+                                // Mise en forme (boutons alignés verticalement
+                                echo '<td class="btn-colonne">';
+                                echo '<div class="d-flex justify-content-center gap-1">';
+                                ?>
+                                <!-- Paramètre envoyé pour supprimer la salle -->
+                                <form method="post" onsubmit="return confirm('Êtes-vous sûr de vouloir supprimer cette salle ?')">
+                                    <input name="idSalle" type="hidden" value="<?php echo $salle['id_salle']; ?>">
+                                    <input name="supprimer" type="hidden" value="true">
+                                    <button type="submit" class="btn-suppr rounded-2"><span class="fa-solid fa-trash"></span></button>
+                                </form>
 
-                            // Mise en forme (boutons alignés verticalement
-                            echo '<td class="btn-colonne">';
-                            echo '<div class="d-flex justify-content-center gap-1">';
-                            ?>
-                            <!-- Paramètre envoyé pour supprimer la salle -->
-                            <form method="post" onsubmit="return confirm('Êtes-vous sûr de vouloir supprimer cette salle ?')">
-                                <input name="idSalle" type="hidden" value="<?php echo $salle['id_salle']; ?>">
-                                <input name="supprimer" type="hidden" value="true">
-                                <button type="submit" class="btn-suppr rounded-2"><span class="fa-solid fa-trash"></span></button>
-                            </form>
-
-                            <!-- Paramètre envoyé pour modifier la salle -->
-                            <form  method="post" action="modificationSalle.php">
-                                <input name="idSalle" type="hidden" value="<?php echo $salle['id_salle']; ?>">
-                                <button type="submit" class="btn-modifier rounded-2"><span class="fa-regular fa-pen-to-square"></span></button>
-                            </form>
-                            <?php
-                            echo '</div>';
-                            echo "</td>";
+                                <!-- Paramètre envoyé pour modifier la salle -->
+                                <form  method="post" action="modificationSalle.php">
+                                    <input name="idSalle" type="hidden" value="<?php echo $salle['id_salle']; ?>">
+                                    <button type="submit" class="btn-modifier rounded-2"><span class="fa-regular fa-pen-to-square"></span></button>
+                                </form>
+                                <?php
+                                echo '</div>';
+                                echo "</td>";
+                            }
                             echo "</tr>";
                         }
                     }
